Move to test-helper
RISDEV-6154
